requirements.txt
pytest
python-rapidjson
pytz
simplejson
tabulate
ujson
yajl
